Raycaster outline update

+++ app/root/shaders/main/outline.vert.glsl
@@ -1,5 +1,5 @@
 void setOutlineVert() {
-    int thickness = 0.05;
+    float thickness = 0.05;
     vec3 worldPos = aPos + aNormal * thickness;
     
     gl_Position = uProjection * uView * uModel * vec4(worldPos, 1.0);
